requirements.txt
# Core data processing
numpy>=1.20.0
pandas>=1.3.0  # Used for efficient CSV processing
scipy>=1.15.0
scikit-learn>=1.6.0

# Audio processing
librosa>=0.11.0
pydub>=0.25.1
soundfile>=0.13.0

# Visualization
matplotlib>=3.5.0
seaborn>=0.11.0

# API clients
deepgram-sdk>=3.1.0
openai>=1.13.0
google-cloud-storage>=3.1.0
convex>=0.7.0

# Hugging Face and ML models for local transcription
transformers>=4.36.0
torch>=2.1.0
torchaudio>=2.1.0
accelerate>=0.25.0

# NeMo toolkit for NVIDIA Parakeet models
nemo_toolkit[asr]>=2.0.0

# Utilities
tqdm>=4.62.0
python-dotenv>=1.1.0
PyYAML>=6.0.0
emoji>=2.14.0
psutil>=7.0.0
ijson>=3.2.3  # Streaming JSON parser
requests>=2.31.0  # HTTP client for Convex API

# Cloud Functions
functions-framework>=3.8.0
google-api-core>=2.24.0
google-auth>=2.39.0

# Async support
aiohttp>=3.11.0

# FastAPI and async web framework
fastapi>=0.104.0
uvicorn>=0.24.0
pydantic>=2.5.0
